requirements.txt
anyio==3.7.0
black==23.3.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.98.0
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pydantic==1.10.9
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
uvicorn==0.22.0
